+++ oceanics-io-ui/src/components/Campaign/PageData.json
@@ -9,6 +9,6 @@
         "callToAction": "The sea is waiting",
         "response": "Take the helm",
         "name": "εἴδωλον",
-        "description": "To protect our Ocean, you need to draw on community knowledge and make data-driven decisions for the future.\nOceanics.io analyzes public and proprietary assets and serves you synthetic and aggregate products to manage risk and conflict.\nWe demystify the sea and amplify capabilities, whether watching your surf or seeking opportunity."
+        "description": "To protect our Ocean, you need to draw on community knowledge and make data-driven decisions for the future.\nWe analyze public and proprietary assets and serve you synthetic and aggregate products to manage risk and conflict. Whether watching your surf or seeking opportunity."
     }]
 }
